MinerAI/requirements.txt

numpy
gymnasium
stable-baselines3[extra]
torch
pygame
opencv-python
pyautogui
selenium
tensorboard
matplotlib